PHPStan L7 全量清零：211 → 0，主配置 level 5 → 7
在 L6 清零基础上，将 phpstan 探测级别提升至 L7，全量修复 211 条真实错误，
并将主配置从 level 5 直接固化到 level 7（phpstan-baseline 保持为空）。

## 功能性修复（真 bug / 崩溃点）

- 支付处理器（13 个文件）：User::find / Plan::find 的 Model|Collection
  联合未窄化，统一改为 query()->where(主键)->first()/firstOrFail()，
  并补 getPrice() 的 float 强转（prices 数组兜底分支返回数组时原本返回非标量）
- Stripe / Paystack / Razorpay：total_amount(decimal:2 cast 为 string) 与 100
  相乘前补 (float)(?? 0) 防御，null 乘法崩溃消除
- PaymentService：payment_total_amount 累计补 (float) cast（decimal 字符串相加）
- AuthController 2FA：session 中 user_id 查询窄化；注册黑名单
  strrchr() false 分支防御（无 @ 域名邮箱原本触发 substr(false) TypeError）
- Heatmap / Goal / Annotation / Domain / SeoAudit / SSO / AdminPayments：
  find()/relation findOrFail() 联合窄化 + 无 null 检查处改 firstOrFail，
  非法 ID 由 500 崩溃变为 404
- AdminTaxes / WebsitesImport CSV 导入：$file->getRealPath() false 分支 +
  fopen false 防御，异常时 422 而非 TypeError
- WebhookEmailController：query token 数组值 is_string 检查（防 (string)array）
- Captcha：widget() provider null 防御（原 null 键访问产出坏 HTML）；
  geetest json_encode false 分支
- SetLocale / SocialLogin Apple JWT / TencentSms / PixelTracker / NotificationDispatcher /
  AuditEngine / LicenseManager：json_encode / hex2bin / gzencode /
  openssl_pkey_derive / openssl_pkey_get_details / openssl_pkey_new /
  curl_exec 等函数 false 分支逐一防御（WebPush 加密管线失败改抛 RuntimeException）
- StatisticsService exitPages：website_id 从 SQL 字符串拼接改为绑定参数
- DemoHeatmapReplaySeeder：ip2long false 分支防御
- QQProvider：getAccessToken 返回值归一化（parse_str out-param 类型闭合）
- curl 客户端 ×3（S3 / AliyunOSS / TencentCOS）：CURLOPT_CUSTOMREQUEST
  空串防御（非空字符串要求）
- SeoKeywordRank：补 @property 注释块（其余模型均已具备）

## 流程性

- phpstan.neon level 5 → 7：CI/本地守护直接运行在最高实测级别，
  baseline 保持为空（0 条），新增错误一律本地暴露
- 部署链路验证：view:cache / config:cache / route:cache 全部通过
  （无 closure 路由、配置可序列化），验证后已 clear 还原
- composer validate 通过；vite build 通过

## 验证矩阵

- PHPStan L7（app/routes/config/database 全路径，空 baseline）：0 错误
- pint 全仓：489 文件 PASS
- 测试：608 passed / 2328 assertions 全绿
- HTTP 冒烟：landing / login 均 200
- 全仓语法扫描：737 个 PHP 文件 0 语法错误
- SQL 注入扫描：raw 系拼接全部为白名单内插或参数绑定，0 风险

65 files changed, 229 insertions(+), 148 deletions(-)

+++ app/Services/PlanLimitService.php
@@ -69,7 +69,7 @@ class PlanLimitService
             return 0; // 未启用
         }
 
-        return max(0, $limit - $this->getCurrentUsage($user, $feature));
+        return (int) max(0, $limit - $this->getCurrentUsage($user, $feature));
     }
 
     /**
